feat: allow downloading items  without manifests
